+++ CHANGELOG.md
@@ -14,6 +14,7 @@ values whitelisted by the `:acceptable_authenticator_statuses` option
 - [Wax.Challenge] Added timeout field and verification
 - [Wax] Added `:android_key_allow_software_enforcement` option
 - [Wax.Metadata] FIDO2 metadata TOC checked against CRLs (#12)
+- [Wax] Added a mean to load metadata from a directory
 
 ### Changed